+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 Status of the `main` branch. Changes prior to the next official version change will appear here.
 
+* JetBrains:
+  - `JetBrainsRenameTool` can now also rename occurrences in comments and text.
+
 * Language Servers:
-  - Fix Dart LSP returning only symbol name as body instead of full method body
+  - Fix Dart LSP returning only symbol name as body instead of full method body.
 
 # 1.1.0