+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 Status of the `main` branch. Changes prior to the next official version change will appear here.
 
+* General:
+  - Enable cert verification for HTTPS request to oraios-software.de #1320
+
 * JetBrains:
   - `JetBrainsRenameTool` can now also rename occurrences in comments and text.
 
@@ -11,8 +14,8 @@ Status of the `main` branch. Changes prior to the next official version change w
 # 1.1.0
 
 * General:
-  - Add `serena init` and `serena setup` commands
   - **Major**: Add commands for hooks and documentation of recommended setup. Consider setting up the [recommended hooks](https://oraios.github.io/serena/02-usage/030_clients.html) !
+  - Add `serena init` and `serena setup` commands
   - Rework installation instructions, switching to releases on pypi for distribution. Please update your mcp startup commands!
   - Add minimal usage data collection on startup (only Serena version, language backend, OS, dashboard enabled status; no personally identifiable information)
   - Fix: git commit id in Serena version strings was incorrect